+++ resharper/resharper-unity/src/Unity/CSharp/Feature/Services/ContextActions/GenerateBakerAndComponentAction.cs
@@ -9,9 +9,10 @@ using JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes;
 using JetBrains.ReSharper.Plugins.Unity.Resources;
 using JetBrains.ReSharper.Plugins.Unity.Resources.Icons;
 using JetBrains.ReSharper.Plugins.Unity.UnityEditorIntegration.Api;
-using JetBrains.ReSharper.Psi;
+using JetBrains.ReSharper.Plugins.Unity.Utils;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.ReSharper.Psi.Tree;
+using JetBrains.ReSharper.Resources.Resources.Icons;
 using JetBrains.Util;
 
 namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.ContextActions
@@ -35,7 +36,8 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.ContextActio
 
             var fix = new GenerateBakerAndComponentActionFix(classDeclaration, node);
             
-            var action = new IntentionAction(fix, UnityGutterIcons.UnityLogo.Id, new SubmenuAnchor(BulbMenuAnchors.PermanentBackgroundItems, SubmenuBehavior.Executable));
+            var action = new IntentionAction(fix, PsiFeaturesUnsortedThemedIcons.FuncZoneGenerate.Id,
+                new SubmenuAnchor(IntentionsAnchors.HighPriorityContextActionsAnchor, SubmenuBehavior.Executable));
 
             return new[] {action};
         }
@@ -52,8 +54,8 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.ContextActio
                 node.GetContainingNode<IPropertyDeclaration>() != null)
                 return false;
             
-            return UnityApi.IsDerivesFromComponent(classDeclaration.DeclaredElement)
-                && myDataProvider.Solution.HasEntitiesPackage();
+            return classDeclaration.DeclaredElement.DerivesFrom(KnownTypes.Component)
+                   && myDataProvider.Solution.HasEntitiesPackage();
 
         }
     }