+++ resharper/resharper-unity/test/data/Unity/CSharp/Intentions/QuickFixes/Dots/DotsMustBeStruct/ScopeTest02.cs
@@ -1,0 +1,10 @@
+using Unity.Entities;
+
+
+public class BotData{caret}Aspect : IAspect
+{
+}
+
+public readonly class IKAnimationAspect : IAspect
+{
+}